Move tos to separate django app
