+++ ecc/settings.py
@@ -63,6 +63,7 @@ INSTALLED_APPS = [
     'utils',
     'adauth',
     'session',
+    'tos',
 ]
 
 
@@ -75,7 +76,7 @@ MIDDLEWARE = [
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'control.middleware.WelcomeMiddleware',
+    'tos.middleware.WelcomeMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.contrib.sites.middleware.CurrentSiteMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
