+++ ecc/urls.py
@@ -16,6 +16,7 @@ from magicauth import views as magicauth_views
 from magicauth.urls import urlpatterns as magicauth_urls
 from user_profiles import api_views as user_profiles_api_views
 from session import api_views as session_api_views
+from tos import views as tos_views
 
 
 admin.site.site_header = 'e-contrôle Administration'
@@ -37,7 +38,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('logout/', auth_views.LogoutView.as_view(next_page='/'), name='logout'),
 
-    path('bienvenue/', control_views.Welcome.as_view(), name='welcome'),
+    path('bienvenue/', tos_views.Welcome.as_view(), name='welcome'),
     path('accueil/', control_views.QuestionnaireList.as_view(), name='questionnaire-list'),
     path('questionnaire/<int:pk>/', control_views.QuestionnaireDetail.as_view(), name='questionnaire-detail'),
     path('questionnaire/controle-<int:pk>/creer',
