+++ tos/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TosConfig(AppConfig):
+    name = 'tos'
